Refactor tree_walk and doc generation
Factor out functions that do walk-tree and that answer questions about the
model (e.g. is_optional()? for a particular field).  Those functions are
reusable for several tasks and should not be tied to document generation
only.

The implementations of helper functions are now also cleaner, and using
consistent naming of parameters.

The ifex_ast_doc.py was modified accordingly to make use of the new
module ifex_model_walk for walk-function and other helpers.
Other than that, very little change to the actual documentation
generation logic.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>