Merge-tree implementation
This is a fundamental implementation that implements/defines the rules for
combining Layers.  It implements the ability to add, delete, and change
information, both in the case of overlays that modify existing interface
definitions within the IFEX Core IDL scope, and new layers that *augment* the
information with new information types.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>